requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.4.0
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.0
